+++ hosts/phi/home.nix
@@ -3,17 +3,7 @@
 let
   inherit (lib) mkForce;
 in {
-  home.packages = builtins.attrValues {
-    inherit (pkgs) signal-desktop;
-  };
-
-  # Disable tab bar when using vertical tabs
-  home.file.".mozilla/firefox/userChrome.css".text = ''
-    #TabsToolbar { visibility: collapse !important; }
-  '';
-
   xsession.windowManager.i3.config.startup = [
-    { command = "signal-desktop"; always = true; }
     { command = "i3 workspace 101"; notification = false; }
     { command = "i3 workspace 201"; notification = false; }
   ];